refactored to a sevice method
